+++ CHANGELOG.md
@@ -19,6 +19,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 
 
+## [0.0.3] - 2025-07-12
+
+### Added
+
+- Tailwind Variants 
+- Tailwind Merge
+- clsx
+
+### Fixed
+
+- Button rendering issue. 
+
+### Changed
+
+- 
+
+### Removed
+
+-
 
 
 ## [0.0.2] - 2025-07-12